split into config file
