--- .env.vault
@@ -1,25 +1,0 @@
-#/-------------------.env.vault---------------------/
-#/         cloud-agnostic vaulting standard         /
-#/   [how it works](https://dotenv.org/env-vault)   /
-#/--------------------------------------------------/
-
-# development
-DOTENV_VAULT_DEVELOPMENT="TdigTpm2AYkoX+n8fqetDxtukDEInZ2an3pQIZV8aD8N59yHfjCL8i4lLWcFa91thbK1+Sm23gc2ZtRiVemkkjJX6/Ym7z6W1snhRbUMPw69jyDoFBKfhmu8/2nTBcEvquurhswCtfAhRBLoeDHFNULP6aP0r6AllQK/tTRtLBYhplqnt+up+w3Nkx6am+iFtyqV3hKZhgFC7G/vh3V/YilB4HtSXGjj8968xIY2yeLjfpCGHqpB+06jaDf+Ypsu3etojDhfRnXRW3V+VfMWnpkxZGP/0Wr79+nOiZ+H8Y1KjXd/v4M66s/lwZVtdBMamivAj4QtyR4QyYwbw1PhtkEnrUmtBoXgGaPnu4sSMUvvIqce9cgGQcrgHQaLUizRFkQXWLaC5D0idQdheq/tQgybsM+UAuCCxEL76RyBwlKp6lhUUHw2jiFYS3u2/5oxtP7kKj4f7YdCILtQvDtsf0GBQGMiuqj0l4jZg7l+6MtT+2qqK3fcbXtYJV4I/GzyYxZw0RfJyzhYDY2Evzx9HxOCIpofQRgfFoJ4qj3gE37XjtGfCy61ScMp8g9uM4RS"
-DOTENV_VAULT_DEVELOPMENT_VERSION=2
-
-# ci
-DOTENV_VAULT_CI="OE9nFgSLDDwUkw0ZiEMtgWc7Q71Sut0RkUy5sR4iywQrjKL2"
-DOTENV_VAULT_CI_VERSION=1
-
-# staging
-DOTENV_VAULT_STAGING="s4/EYP6noF1qduKK4hwBrTKOC0ZUyvPPFUyw+aremdCMlS2C"
-DOTENV_VAULT_STAGING_VERSION=1
-
-# production
-DOTENV_VAULT_PRODUCTION="bPMDNCdlQ+mrNkfbXzmjpxsLFf4opZU7m3MdLyrvuFSP79sN"
-DOTENV_VAULT_PRODUCTION_VERSION=1
-
-#/----------------settings/metadata-----------------/
-DOTENV_VAULT="vlt_bbdd92e351c43dcfac195bf872e8a6c495ca7cf6bb588eea5d73a0c61d8fb7fc"
-DOTENV_API_URL="https://vault.dotenv.org"
-DOTENV_CLI="npx dotenv-vault@latest"
