+++ src/app/layout.tsx
@@ -1,16 +1,13 @@
-import React from 'react'
+import { Toaster } from '@/components/ui/toaster'
+import { GoogleAnalytics } from '@next/third-parties/google'
+import _isEmpty from 'lodash/isEmpty'
+import type { Metadata, Viewport } from 'next'
 import Head from 'next/head'
 import Link from 'next/link'
-import icons from '@/metadata/icons'
-import _isEmpty from 'lodash/isEmpty'
-import robots from '@/metadata/robot'
-import twitter from '@/metadata/twitter'
-import openGraph from '@/metadata/open-graph'
-import type { Metadata, Viewport } from 'next'
-import { GoogleAnalytics } from '@next/third-parties/google'
-import { Toaster } from '@/components/ui/toaster'
+import React, { PropsWithChildren } from 'react'
 
 import '@/app/globals.css'
+import metadataConfig from '@/config/metadata'
 
 const { GOOGLE_ANALYTICS_ID } = process.env
 
@@ -18,30 +15,9 @@ export const viewport: Viewport = {
   themeColor: '#3A86FF'
 }
 
-export const metadata: Metadata = {
-  title: 'StudyCrew',
-  description: 'Making education more accessible, collaborative, and engaging.',
-  applicationName: 'StudyCrew',
-  manifest: '/manifest.json',
-  icons,
-  appleWebApp: {
-    capable: true,
-    statusBarStyle: 'default',
-    title: 'StudyCrew'
-  },
-  formatDetection: {
-    telephone: false
-  },
-  openGraph,
-  robots,
-  twitter
-}
+export const metadata: Metadata = metadataConfig
 
-interface LayoutProps {
-  children: React.ReactNode
-}
-
-const RootLayout = ({ children }: LayoutProps): JSX.Element => {
+export default function RootLayout({ children }: PropsWithChildren) {
   return (
     <html lang="en">
       <Head>
@@ -58,5 +34,3 @@ const RootLayout = ({ children }: LayoutProps): JSX.Element => {
     </html>
   )
 }
-
-export default RootLayout
