+++ src/app/robots.ts
@@ -4,7 +4,8 @@ export default function robots(): MetadataRoute.Robots {
   return {
     rules: {
       userAgent: '*',
-      allow: '/'
+      allow: ['/'],
+      disallow: ['/dashboard', '/groups', '/profile']
     },
     sitemap: 'https://studycrew.world/sitemap.xml'
   }
