+++ src/config/site.ts
@@ -1,0 +1,9 @@
+const siteConfig = {
+  title: 'StudyCrew',
+  shortName: 'studycrew',
+  description: 'Making education more accessible, collaborative, and engaging.',
+  applicationName: 'StudyCrew',
+  manifest: '/manifest.json'
+}
+
+export default siteConfig
