--- src/metadata/twitter.ts
@@ -1,12 +1,0 @@
-import { type Twitter } from 'next/dist/lib/metadata/types/twitter-types'
-
-const twitter: Twitter = {
-  card: 'summary',
-  title: {
-    default: 'StudyCrew',
-    template: '% - PWA App'
-  },
-  description: 'Making education more accessible, collaborative, and engaging.'
-}
-
-export default twitter
